requirements.txt
agents
agents-arcade
arcadepy
libsql-experimental
openai
python-dotenv
pydantic
